refactor: move inbox hook tests to dedicated inbox-hooks.test.ts
Extracts lifecycle hook tests from realtime.test.ts into their own
file since they test hook behavior, not realtime publishing. Adds a
test verifying that a throwing hook does not block the realtime
publish.
